refactor: extract places feature to feature-places module
Moves BackupTool to data:files and SendIntentResolver to ui-common
first, since both were misplaced inside feature-reminder/feature-settings
with no feature-specific logic, which blocked feature-places from
using them under the "features never depend on each other" rule.
EditPlaceViewModel now reads marker prefs through ui-map's
MapPreferences seam instead of app's concrete Prefs.
