+++ app/build.gradle.kts
@@ -232,6 +232,7 @@ dependencies {
   implementation(project(":feature:feature-insights"))
   implementation(project(":feature:feature-workflow"))
   implementation(project(":feature:feature-settings"))
+  implementation(project(":feature:feature-places"))
   implementation(project(":extensions:localbackup"))
 
   implementation(project(":core:platform-common"))
