+++ data/files/src/main/kotlin/com/github/naz013/files/KoinModule.kt
@@ -1,8 +1,10 @@
 package com.github.naz013.files
 
+import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
 
 val fileModule = module {
   factory { DataConverterImpl() as DataConverter }
   factory { AndroidDataConverterImpl(get(), get()) as AndroidDataConverter }
+  singleOf(::BackupTool)
 }
