+++ feature/feature-places/build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
 }
 
 android {
-  namespace = "com.github.naz013.feature.reminder"
+  namespace = "com.github.naz013.feature.places"
 
   testOptions {
     unitTests {
@@ -15,55 +15,28 @@ android {
 
 dependencies {
   implementation(project(":core:domain"))
+  implementation(project(":data:repository-api"))
   implementation(project(":core:logging-api"))
   implementation(project(":core:navigation-api"))
-  implementation(project(":core:analytics"))
   implementation(project(":core:platform-common"))
-  implementation(project(":core:platform-api"))
   implementation(project(":core:date-calculations"))
   implementation(project(":core:feature-common"))
-  implementation(project(":core:feature-flags-api"))
-  implementation(project(":data:repository-api"))
   implementation(project(":data:files-api"))
   implementation(project(":data:files"))
-  implementation(project(":data:cloud-api"))
-  implementation(project(":data:icalendar-api"))
-  implementation(project(":data:googlecalendar-api"))
-  implementation(project(":data:work-api"))
-  implementation(project(":data:scheduler-api"))
-  implementation(project(":extensions:appwidgets-api"))
-  implementation(project(":admin:reviews"))
-  implementation(project(":ui:ui-common"))
-  implementation(project(":ui:ui-reminder"))
-  implementation(project(":ui:ui-note"))
-  implementation(project(":ui:ui-tag"))
-  implementation(project(":ui:ui-group"))
-  implementation(project(":ui:ui-googletask"))
-  implementation(project(":ui:ui-map"))
-  implementation(project(":ui:ui-notification-settings"))
-  implementation(project(":logic:logic-reminder"))
   implementation(project(":logic:logic-schedule"))
-  implementation(project(":logic:logic-tag"))
-  implementation(project(":logic:logic-workflow"))
-  implementation(project(":feature:feature-tags"))
+  implementation(project(":ui:ui-common"))
+  implementation(project(":ui:ui-map"))
 
   implementation(libs.koin.android)
   implementation(libs.koin.android.ext)
   implementation(libs.koin.androidx.compose)
   implementation(libs.koin.compose.navigation3)
-  implementation(libs.threetenbp)
-  implementation(libs.gson)
 
   implementation(libs.androidx.navigation3.runtime)
   implementation(libs.androidx.navigation3.ui)
   implementation(libs.androidx.lifecycle.viewmodel.navigation3)
   implementation(libs.kotlinx.serialization.core)
-
-  implementation(libs.coil)
-  implementation(libs.coil.compose)
-
-  implementation(libs.lottie)
-  implementation(libs.lottie.compose)
+  implementation(libs.threetenbp)
 
   implementation(platform(libs.compose.bom))
   implementation(libs.compose.material3)
@@ -72,7 +45,6 @@ dependencies {
   implementation(libs.compose.ui.ui)
   implementation(libs.compose.ui.tooling.preview)
   implementation(libs.compose.runtime.livedata)
-
   debugImplementation(libs.compose.ui.tooling)
 
   testImplementation(project(":core:testing"))
