+++ feature/feature-places/src/main/kotlin/com/github/naz013/feature/places/KoinModule.kt
@@ -1,15 +1,15 @@
-package com.elementary.tasks.places
+package com.github.naz013.feature.places
 
-import com.elementary.tasks.places.create.EditPlaceViewModel
-import com.elementary.tasks.places.list.PlacesViewModel
-import com.elementary.tasks.places.usecase.DeletePlaceUseCase
-import com.elementary.tasks.places.usecase.SavePlaceUseCase
+import com.github.naz013.feature.places.create.EditPlaceViewModel
+import com.github.naz013.feature.places.list.PlacesViewModel
+import com.github.naz013.feature.places.usecase.DeletePlaceUseCase
+import com.github.naz013.feature.places.usecase.SavePlaceUseCase
 import org.koin.core.module.dsl.factoryOf
 import org.koin.core.module.dsl.viewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
-val placeKoinModule = module {
+val featurePlacesModule = module {
   factoryOf(::DeletePlaceUseCase)
   factoryOf(::SavePlaceUseCase)
