+++ feature/feature-places/src/main/kotlin/com/github/naz013/feature/places/create/EditPlaceScreen.kt
@@ -1,4 +1,4 @@
-package com.elementary.tasks.places.create
+package com.github.naz013.feature.places.create
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -22,7 +22,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.elementary.tasks.R
+import com.github.naz013.ui.common.R
 import com.github.naz013.ui.common.compose.AppIcons
 import com.github.naz013.ui.common.compose.foundation.MenuIconButton
 import com.github.naz013.ui.common.compose.foundation.MenuTextButton
