+++ feature/feature-places/src/test/kotlin/com/github/naz013/feature/places/list/PlacesViewModelTest.kt
@@ -1,10 +1,10 @@
-package com.elementary.tasks.places.list
+package com.github.naz013.feature.places.list
 
-import com.elementary.tasks.BaseTest
-import com.elementary.tasks.R
-import com.github.naz013.feature.reminder.util.BackupTool
-import com.elementary.tasks.mockDispatcherProvider
-import com.elementary.tasks.places.usecase.DeletePlaceUseCase
+import com.github.naz013.testing.BaseTest
+import com.github.naz013.ui.common.R
+import com.github.naz013.files.BackupTool
+import com.github.naz013.testing.mockDispatcherProvider
+import com.github.naz013.feature.places.usecase.DeletePlaceUseCase
 import com.github.naz013.ui.map.MapStyle
 import com.github.naz013.common.TextProvider
 import com.github.naz013.datecalc.DateTimeManager
