+++ feature/feature-reminder/src/main/kotlin/com/github/naz013/feature/reminder/KoinModule.kt
@@ -78,7 +78,6 @@ import com.github.naz013.feature.reminder.preview.PreviewReminderViewModel
 import com.github.naz013.feature.reminder.todo.TodoEditNavKey
 import com.github.naz013.feature.reminder.todo.TodoEditViewModel
 import com.github.naz013.feature.reminder.todo.TodoSeedHolder
-import com.github.naz013.feature.reminder.util.BackupTool
 import com.github.naz013.feature.reminder.util.UriHelper
 import com.github.naz013.ui.reminder.UiReminderListAdapter
 import org.koin.core.module.dsl.factoryOf
@@ -343,6 +342,5 @@ val featureReminderModule = module {
   factory { UiReminderCommonAdapter(get(), get(), get(), get(), get(), get(), get()) }
 
   factoryOf(::UriHelper)
-  singleOf(::BackupTool)
   factoryOf(::PresetInitProcessor)
 }
