+++ settings.gradle.kts
@@ -93,6 +93,7 @@ include(":feature:feature-tags")
 include(":feature:feature-insights")
 include(":feature:feature-workflow")
 include(":feature:feature-settings")
+include(":feature:feature-places")
 
 // extensions - cross-feature, flavor/runtime-gated additions app pulls in as a unit
 include(":extensions:appwidgets")
